Extract SttpUpickleApiExtension for better readability
