+++ core/src/test/scala/sttp/openai/requests/embeddings/EmbeddingsDataSpec.scala
@@ -4,7 +4,7 @@ import org.scalatest.EitherValues
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import sttp.openai.fixtures
-import sttp.openai.json.SttpUpickleApiExtension
+import sttp.openai.json.SttpOpenAIApi
 import EmbeddingsResponseBody._
 import sttp.openai.requests.embeddings.EmbeddingsRequestBody.EmbeddingsModel
 
@@ -31,7 +31,7 @@ class EmbeddingsDataSpec extends AnyFlatSpec with Matchers with EitherValues {
     )
     // when
     val givenResponse: Either[Exception, EmbeddingResponse] =
-      SttpUpickleApiExtension.deserializeJsonSnake[EmbeddingResponse].apply(listFilesResponse)
+      SttpOpenAIApi.deserializeJsonSnake[EmbeddingResponse].apply(listFilesResponse)
 
     // then
     givenResponse.value shouldBe expectedResponse
